+++ R/get_label.R
@@ -12,7 +12,8 @@
 #'
 #' @seealso See package vignettes or \href{http://www.strengejacke.de/sjPlot/}{online documentation}
 #'            for more details; \code{\link{set_label}} to manually set variable labels or \code{\link{get_labels}}
-#'            to get value labels.
+#'            to get value labels; \code{\link{var_labels}} to set multiple variable
+#'            labels at once.
 
 #' @param x \code{data.frame} with variables that have label attributes (e.g.
 #'          from an imported SPSS, SAS or STATA data set, via \code{\link{read_spss}},
@@ -33,7 +34,11 @@
 #'
 #' @details See 'Details' in \code{\link{get_labels}}.
 #'
-#' @note See 'Note' in \code{\link{get_labels}}.
+#' @note \code{\link{var_labels}} is an alternative way to set variable labels,
+#'       which follows the philosophy of tidyvers API design (data as first argument,
+#'       dots as value pairs indicating variables).
+#'       \cr \cr
+#'       Furthermore, see 'Note' in \code{\link{get_labels}}.
 #'
 #' @examples
 #' # import SPSS data set